refactor: split engine into modules
