+++ flujo/application/__init__.py
@@ -2,7 +2,7 @@
 Application-level components for flujo.
 """
 
-from .flujo_engine import Flujo
+from .runner import Flujo
 from .eval_adapter import run_pipeline_async
 from .self_improvement import evaluate_and_improve, SelfImprovementAgent
